backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==22.10.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
channels==4.0.0
channels-auth-token-middlewares==1.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.1
daphne==4.0.0
defusedxml==0.8.0rc2
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==4.2.15
django-allauth==0.57.0
django-cors-headers==4.4.0
django-oauth-toolkit==2.4.0
django-websockets==0.0.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-social-oauth2==3.1.0
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.65.4
gunicorn==23.0.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.0
jwcrypto==1.5.6
msgpack==1.0.8
oauthlib==3.2.2
odict==1.9.0
packaging==24.2
pillow==10.4.0
plumber==1.7
proto-plus==1.24.0
protobuf==5.27.3
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.1.3
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
service-identity==24.1.0
setuptools==75.8.0
shortuuid==1.0.13
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
websockets==11.0.1
whitenoise==6.7.0
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==7.0
zope.interface==7.0.1
zope.lifecycleevent==5.0
zope.proxy==6.1
